+++ hooks/Event/useCreateEvent.ts
@@ -1,7 +1,6 @@
-import { createEvent, getEventsList } from '@/lib/supabase/apis/event';
+import { createEvent } from '@/lib/supabase/apis/event';
 import { CreateEventType } from '@/lib/supabase/apis/event/type';
-import { Event } from '@/lib/supabase/type';
-import { useMutation, useQuery, useQueryClient, UseQueryOptions } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 export const EVENT_KEY = {
   all: ['event'],
@@ -22,16 +21,5 @@ export const useCreateEvent = () => {
         throw new Error(err.message);
       },
     }
-  );
-};
-
-export const useFetchEventList = (options?: UseQueryOptions<Event[], Error>) => {
-  return useQuery<Event[], Error>(
-    EVENT_KEY.all,
-    async () => {
-      const response = await getEventsList();
-      return response;
-    },
-    { ...options }
   );
 };
